Rename handler.rs to session.rs
