+++ src/lib.rs
@@ -5,7 +5,7 @@
 
 pub mod config;
 pub mod error;
-pub mod handler;
+pub mod session;
 pub mod input;
 pub mod keymap;
 pub mod shell;
@@ -14,7 +14,7 @@ pub mod ui;
 
 pub use config::Config;
 pub use error::LeadrError;
-pub use handler::LeadrSession;
+pub use session::LeadrSession;
 pub use shell::init_bash;
 pub use shell::init_zsh;
 pub use types::{Mapping, LeadrResult, Mappings};
